move displaying products to ProductsList component, fetch products on sale in Deals component
