+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 import './App.css';
 import Navbar from './components/Navbar'
-import Home from './components/home/Home'
+import Home from './components/Home/Home'
 import Cart from './components/Cart'
 import Deals from './components/Deals'
