+++ src/components/Home/Home.css
@@ -1,0 +1,6 @@
+/* .items-list {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(320px, 1fr));
+    grid-gap: 1rem;
+    justify-items: center;
+} */
