+++ src/components/Item/Item.css
@@ -2,6 +2,8 @@
     border: 1px solid black;
     height: 320px;
     width: 300px;
+    display: grid;
+    
 }
 
 .item-img {
