+++ src/components/Stars/EmptyStar.js
