+++ src/components/Stars/FullStar.js
