+++ src/components/Stars/HalfStar.js
